consensus: Add consensus peers health reporting test
This new test provides functional coverage on tablet health states and
common health status transitions.

This patch also adds a version of CauseFollowerToFallBehindLogGC() that
allows for a pre-selected follower to fall behind the log GC threshold.
That enables us to verify a health status transition from FAILED to
FAILED_UNAVAILABLE.

Change-Id: I65d787a3146749e36e33d78e253914a5ffe5762d
Reviewed-on: http://gerrit.cloudera.org:8080/9766
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>