design-docs: improve cfile.md
I've continually found the CFile abstraction difficult to place in
context with the higher-level DiskRowSet and lower-level BlockManager
abstractions. Every time I revisit the CFile code I have to re-research
these relationships and interactions. cfile.md is currently lacking in
these details, so this is an attempt to fix that.

Change-Id: I770028bba3f7a49c96f32893c285221c84be39ce
Reviewed-on: http://gerrit.cloudera.org:8080/8860
Tested-by: Kudu Jenkins
Reviewed-by: David Ribeiro Alves <davidralves@gmail.com>