KUDU-2236: fix flake in TestKuduClient
TestKuduClient.testCloseShortlyAfterOpen would expect 0 log spew, and
ensure that no unexpected exceptions were caught. It noted that
"connection disconnected" errors were possible, but added a wait to make
them less likely. This turned out to be extremely flaky.

This patch replaces the wait with logic to catch the appropriate error
message.

Without this patch, the test only passed 57/100 runs. With the patch, it
passes 100/100.

Change-Id: Ice5b404107dd425eab5ea3f382090b2304ef8a5e
Reviewed-on: http://gerrit.cloudera.org:8080/8892
Reviewed-by: Hao Hao <hao.hao@cloudera.com>
Tested-by: Kudu Jenkins