data_dirs: fix logging message
If fs_target_data_dirs_per_tablet is set to be greater than the number
of available directories, it will log a message that is dependent on
there being a configured metrics entity, which is not always available.
This patch avoids the potential nullptr access.

This patch also changes the logging to INFO-level instead of
WARNING-level, as the message doesn't necessarily indicate a problem.

Change-Id: I8135e0ad2411524fcb74b3e349bea7a5828a4237
Reviewed-on: http://gerrit.cloudera.org:8080/8999
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins