KUDU-2148: do not crash on GetStatus during server startup
The fix is straight-forward but there are backwards compatibility
implications for old clients who trigger the race: because they're not aware
of the new 'error' field in the protobuf, they'll just see the incomplete
'status' field. To be more specific, 'node_instance' will be set,
'version_info' won't be, and 'bound_{rpc,http}_addresses' may or may not be
set depending on exactly when the race was hit.

An alternative would be to fail the RPC itself in lieu of adding the 'error'
field. That would ensure that old clients register a failure, though it
comes at the expense of not providing rich error information for new
clients, which is why I went with the above solution instead.

Change-Id: I3b997999d8a1228244e0be0ea61705f2c12e3426
Reviewed-on: http://gerrit.cloudera.org:8080/9041
Reviewed-by: Dan Burkert <danburkert@apache.org>
Tested-by: Kudu Jenkins