[java] add test to reproduce KUDU-2267
Currently, if a master has never been a leader from the very start of
the cluster, it has just self-signed server TLS certificate. In that
case if a client does not have any valid Kerberos credentials but only
authn token, then the client may see negotiation failure error and is
not able to connect. This patch adds a regression test for this issue.

Change-Id: I4879749988dc884fe81cf36838819379db91ae72
Reviewed-on: http://gerrit.cloudera.org:8080/9071
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>