[tests] de-flaking DontEvictIfRemainingConfigIsUnstable
Tuned configuration of the test cluster to minimize chances of reporting
on failed tablet replicas 'one-by-one' instead of 'two-at-once' for the
DontEvictIfRemainingConfigIsUnstableITest-derived scenarios.  The
'one-by-one' reporting on failed replicas was the reason behind
flakiness in the following scenarios:

  * DontEvictIfRemainingConfigIsUnstableITest.NodesDown/{0,1}
  * DontEvictIfRemainingConfigIsUnstableITest.NodesStopped/{0,1}

That was because the logic of those test scenarios assumes 2 out of 3
replicas becoming unresponsive at once.  If only 1 out of 3 replicas
became unresponsive and reported as such, the re-replication logic
(both 3-2-3 and 3-4-3 schemes) tried to initiate replacement of the
failed replica.  That's completely legit since at the time of the single
replica failure report the majority of replicas was reported as being
online.

As an additional fix, the ASSERT_EVENTUALLY() wrapper is added for the
invocation of GetTabletToReplicaUUIDsMapping().  That's to guarantee
that all relevant tablet servers are included into the mapping being
built.

Change-Id: Icb23d71b600a46f7fd6bba56091f884eae24784f
Reviewed-on: http://gerrit.cloudera.org:8080/9126
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>