env: properly use fdatasync on Linux
A previous commit attempted to make env OSX-compatible by checking
whether fdatasync was defined and using fsync in the case that it was
not. However, that commit was buggy since it used #ifndef. fdatasync is
never a preprocessor macro, so it never was "defined" and we always
ended up using fsync regardless of the FLAGS_env_use_fsync value.

Change-Id: I366df37670ec241cbee6dcfe0efa56af7189e9a2
Reviewed-on: http://gerrit.cloudera.org:8080/9128
Reviewed-by: Dan Burkert <dan@cloudera.com>
Tested-by: Kudu Jenkins