build-support: fix reporting of build configuration
Our older Jenkins setup for running flaky test builds seemed to use the
deprecated "build from the source root" setup with cmake. Now, we use a
build dir which isn't the same as the source root, which means we no
longer have a CMakeCache.txt in the source root.

This updates the test reporting to look for CMakeCache.txt in the proper
spot.

Change-Id: Ib6c131f46e75d2eb57d7e0ce9d2ad21e2ad9ab80
Reviewed-on: http://gerrit.cloudera.org:8080/9182
Tested-by: Kudu Jenkins
Reviewed-by: Hao Hao <hao.hao@cloudera.com>
Reviewed-by: Grant Henke <granthenke@gmail.com>