test: deflake TestRestartWithDifferentUUID
The test tried to verify the eventual re-replication of failed tablets.
This wouldn't always work because, given the large number of tablets per
tablet server, the servers would be too busy with handling delete tablet
RPCs to begin re-replicating replicas. This led to a very flaky test
(75/100 failures in TSAN).

This patch reduces the number of tablets and updates the number of
tablet servers to ensure tablet copies can begin in a timely manner on
the restarted tablet server. This passed 1000/1000 in TSAN.

Change-Id: Ice109fe9073e53f1651b30dced200f2cf12e7249
Reviewed-on: http://gerrit.cloudera.org:8080/9190
Tested-by: Kudu Jenkins
Reviewed-by: David Ribeiro Alves <davidralves@gmail.com>