tablet: don't store row count in delta tracker
DeltaTracker's constructor previously required a row-count that can
only be obtained by first reading from disk. While useful as a sanity
check while tracking updates, it's not important for this count to be
known at instantiation-time.

This patch is helpful in reducing the amount of data required from disk
at startup, as one of the reasons we need to fully open some CFiles is
to get this row count. This patch itself doesn't defer any reads; it
just defers the requirement of the row count.

This patch has no functional changes.

Change-Id: I084e0944f388c22e838b017663a812b0ba77245d
Reviewed-on: http://gerrit.cloudera.org:8080/9216
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>