Fix crash when log_dir is not specified
This fixes the tpch1 benchmark after the recent enabling of metrics
logging. The crash was due to FLAGS_log_dir being unset. In typical
servers, we rely on InitGoogleLoggingSafe to default log_dir to /tmp,
but our tpch1 benchmark embeds a client, and a constructor function in
the client calls InitGoogleLoggingSafeBasic which doesn't do this
defaulting. So, we ended up starting the minicluster with an empty log
dir.

This patch just makes it not start the metrics logger if no log dir is
set.

Change-Id: I725a7da038913ff05227c6d5555c34167138148b
Reviewed-on: http://gerrit.cloudera.org:8080/9233
Reviewed-by: Jean-Daniel Cryans <jdcryans@apache.org>
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>