[build] fix on ld's -l:path_to_file notation on OS X
Fixed the sized-deallocation test on OS X.

The -l:<path_to_library> works fine with Linux linker, but this notation
is not supported by ld on OS X.  I didn't notice that because I built
on OS X with OUTPUT_QUIET and ERROR_QUIET options already set
for execute_process(), after verifying that it works as need at Linux.

Now, for Kudu builds on OS X 10.12.x and higher the sized deallocation
is enabled, and on OS X 10.11.x it's disabled.

This is a follow-up for c1b9f1088f830235f9d4e6830f59b1a200a65ab6.

Change-Id: I65304c2d7c37abdddb8c60eac0e288b546cfc530
Reviewed-on: http://gerrit.cloudera.org:8080/9242
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>